common: Handle empty `lib_paths` gracefully
`env::split_paths()` called with an empty string, yields one empty
string. This behavior is not really compatible with the forloop
it's used in, which in such case is going to produce ` -L ` not
followed by any path. This causes rustc to fail.

To prevent this issue, enter the loop only when `lib_paths` is not
empty.

That issue was reproducable on MacOS, where `lib_paths` often ends
up being empty.